+++ src/parse/pipeline.c
@@ -12,6 +12,54 @@
 
 #include "minishell.h"
 
+bool	is_control_operator(t_token *tok)
+{
+	static char *const	operators[] = {"||", "&", "&&", ";", ";;", \
+													"(", ")", "|", "\n"};
+	size_t				i;
+
+	i = 0;
+	while (i < sizeof(operators) / sizeof(*operators))
+	{
+		if (startswith(tok->word, operators[i]))
+			return (TRUE);
+		i++;
+	}
+	return (FALSE);
+}
+
+void	append_command_element(t_node *cmd, t_token **rest, t_token *tok)
+{
+	if (tok->kind == TK_WORD)
+	{
+		append_tok(&(cmd->args), tokdup(tok));
+		tok = tok->next;
+	}
+	else if (equal_op(tok, "<<") && tok->next->kind == TK_WORD)
+		append_node(&(cmd->redirects), redirect_heredoc(&tok, tok));
+	else if (equal_op(tok, "<") && tok->next->kind == TK_WORD)
+		append_node(&(cmd->redirects), redirect_in(&tok, tok));
+	else if (equal_op(tok, ">>") && tok->next->kind == TK_WORD)
+		append_node(&(cmd->redirects), redirect_append(&tok, tok));
+	else if (equal_op(tok, ">") && tok->next->kind == TK_WORD)
+		append_node(&(cmd->redirects), redirect_out(&tok, tok));
+	else
+		parse_error("append_command_element", &tok, tok);
+	*rest = tok;
+}
+
+t_node	*simple_command(t_token **rest, t_token *tok)
+{
+	t_node	*node;
+
+	node = new_node(ND_SIMPLE_CMD);
+	append_command_element(node, &tok, tok);
+	while (tok && !at_eof(tok) && !is_control_operator(tok))
+		append_command_element(node, &tok, tok);
+	*rest = tok;
+	return (node);
+}
+
 t_node	*pipeline(t_token **rest, t_token *tok)
 {
 	t_node	*node;
